Add Optional tablePrefix to InDbTimestampBoundStore constructor
